Added React Boostrap and moved Components and pages into seperate folders so they can have seperate CSS
